SmartDietPlanner/requirements.txt
# ============================================================================
# Smart Diet & Lifestyle Planner - Requirements
# Python Package Dependencies
# ============================================================================

# ============================================================================
# Database
# ============================================================================
psycopg2-binary==2.9.9

# ============================================================================
# GUI Framework
# ============================================================================
# Tkinter comes with Python, no need to install separately

# ============================================================================
# Data Processing and Analysis
# ============================================================================
numpy==1.24.3
pandas==2.0.3

# ============================================================================
# Visualization
# ============================================================================
matplotlib==3.7.2
seaborn==0.12.2

# ============================================================================
# AI and Machine Learning
# ============================================================================
# scikit-learn==1.3.0
# scikit-fuzzy==0.4.2

# ============================================================================
# Natural Language Processing (Optional)
# ============================================================================
# transformers==4.30.2
# openai==0.27.8

# ============================================================================
# Utilities
# ============================================================================
python-dateutil==2.8.2
python-dotenv==1.0.0

# ============================================================================
# Password Hashing
# ============================================================================
bcrypt==4.0.1

# ============================================================================
# Validation
# ============================================================================
email-validator==2.0.0

# ============================================================================
# Export and Reporting
# ============================================================================
reportlab==4.0.4
openpyxl==3.1.2

# ============================================================================
# Configuration Management
# ============================================================================
pyyaml==6.0.1

# ============================================================================
# Testing (Development)
# ============================================================================
pytest==7.4.0
pytest-cov==4.1.0

# ============================================================================
# Code Quality (Development)
# ============================================================================
pylint==2.17.4
black==23.7.0

# ============================================================================
# API (if needed for external integrations)
# ============================================================================
requests==2.31.0

# ============================================================================
# Notes:
# - Install using: pip install -r requirements.txt
# - For development, install test and code quality tools
# - Some packages are commented out - uncomment if needed
# - Update versions as needed for your specific requirements
# ============================================================================
